API bill

+++ back-end/src/main/java/com/vecfonds/backend/payload/request/dto/ShoppingCartDTO.java
@@ -2,13 +2,12 @@ package com.vecfonds.backend.payload.request.dto;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
 @Data
 public class ShoppingCartDTO {
     private Long id;
-    private BigDecimal total = BigDecimal.valueOf(0);
+    private Double total = 0.0;
     private List<CartItemDTO> cartItems = new ArrayList<>();
 }
